# threadpilot/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThreadPilot AI API",
    "description": "MySQL-grounded intent workflow, incremental synchronization and read-only SQL Agent.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development; replace with your team server when deployed"
    }
  ],
  "paths": {
...
    "/chat": {
      "post": {
        "tags": [
          "Workflow"
        ],
        "summary": "Workflow Chat",
        "description": "场景 1.1–3.3：JSON 工作流入口，传回 session_id 延续上下文。",
        "operationId": "workflow_chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workflow/replies": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Receive Reply",
        "description": "场景 3.3：已认证上游回复事件入库；聊天消息不能伪造已收到回复。",
        "operationId": "receive_reply_api_v1_workflow_replies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Receive Reply Api V1 Workflow Replies Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/briefings/{day}/email": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Generate Email",
        "operationId": "generate_email_api_dashboard_briefings__day__email_post",
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Save Email",
        "operationId": "save_email_api_dashboard_briefings__day__email_put",
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Day"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/tasks": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Add Task",
        "operationId": "add_task_api_dashboard_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/tasks/{task_id}": {
      "put": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Update Task",
        "operationId": "update_task_api_dashboard_tasks__task_id__put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/predictions/{order_id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Prediction Source",
        "operationId": "prediction_source_api_predictions__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/sources/{kind}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Diagnosis Source",
        "operationId": "diagnosis_source_api_dashboard_sources__kind__get",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 1500,
            "minLength": 1,
            "title": "Message"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9-]{36}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "confirmation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Id"
          },
          "selected_order_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^ORD-\\d{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Order Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "intent": {
            "$ref": "#/components/schemas/Intent"
          },
          "slots": {
            "$ref": "#/components/schemas/Slots"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "needs_clarification": {
            "type": "boolean",
            "title": "Needs Clarification",
            "default": false
          },
          "clarification_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarification Question"
          },
          "confirmation_required": {
            "type": "boolean",
            "title": "Confirmation Required",
            "default": false
          },
          "confirmation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Id"
          },
          "state": {
            "$ref": "#/components/schemas/DialogState"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order Ids"
          },
          "selected_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Order Id"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "gpt-4.1"
          },
          "business_date": {
            "type": "string",
            "title": "Business Date"
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "intent",
          "slots",
          "confidence",
          "state",
          "business_date"
        ],
        "title": "ChatResponse"
      },
      "DialogState": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "active_order": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Order"
          },
          "active_customer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Customer"
          },
          "pending_clarification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingClarification"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Intent"
              },
              {
                "type": "null"
              }
            ]
          },
          "slots": {
            "$ref": "#/components/schemas/Slots"
          },
          "comparison_orders": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Comparison Orders"
          },
          "last_order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Last Order Ids"
          },
          "draft": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "history": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "History"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DialogState"
      },
      "EmailUpdate": {
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Body"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "body",
          "version"
        ],
        "title": "EmailUpdate"
      },
      "Evidence": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "record": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Record"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "original_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "row",
          "url",
          "record"
        ],
        "title": "Evidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Intent": {
        "type": "string",
        "enum": [
          "order.lookup",
          "order.refresh",
          "order.risk",
          "order.compare",
          "order.commitment",
          "order.prioritize",
          "operations.normality",
          "operations.deviation",
          "execution.chase",
          "execution.note",
          "execution.reminder",
          "unknown"
        ],
        "title": "Intent"
      },
      "NewTask": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "NewTask"
      },
...
      "ReplyEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Event Id"
          },
          "order_id": {
            "type": "string",
            "pattern": "^ORD-\\d{3}$",
            "title": "Order Id"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "order_id",
          "received_at"
        ],
        "title": "ReplyEvent"
      },
...
      "TaskUpdate": {
        "properties": {
          "done": {
            "type": "boolean",
            "title": "Done"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "done",
          "version"
        ],
        "title": "TaskUpdate"
      },
      "ToolCall": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "idempotent": {
            "type": "boolean",
            "title": "Idempotent",
            "default": true
          },
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation",
            "default": false
          },
          "executed": {
            "type": "boolean",
            "title": "Executed",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ToolCall"
      },
...
    }
  }
}